[AN] refactor: main 패키지 이동 #384
