+++ android/app/src/main/AndroidManifest.xml
@@ -45,7 +45,7 @@
             android:name=".presentation.detail.PlayerDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.MainActivity"
+            android:name=".presentation.main.MainActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustPan" />
